Refactor api into own file and added template url creation function
