+++ inventree_bulk_plugin/frontend/src/utils/api.ts
@@ -1,0 +1,30 @@
+export const fetchAPI = (input: RequestInfo | URL, init?: RequestInit) => {
+  return fetch(input, {
+    headers: {
+      "Content-Type": "application/json",
+      ...init?.headers,
+    },
+    ...init,
+  });
+};
+
+export const URLS = {
+  bulkcreate: ({ parentId, create }: { parentId?: string; create?: boolean } = {}) => {
+    const params = new URLSearchParams();
+    if (parentId) params.set("parent_id", parentId);
+    if (create) params.set("create", create ? "true" : "false");
+    const paramsString = params.toString();
+
+    return `/plugin/inventree-bulk-plugin/bulkcreate${paramsString ? `?${paramsString}` : ""}`;
+  },
+  templates: ({ id, templateType }: { id?: number | null; templateType?: string } = {}) => {
+    const params = new URLSearchParams();
+    if (templateType) params.set("template_type", templateType);
+    const paramsString = params.toString();
+
+    return (
+      `/plugin/inventree-bulk-plugin/templates${id !== undefined && id !== null ? `/${id}` : ""}` +
+      `${paramsString ? `?${paramsString}` : ""}`
+    );
+  },
+};
